% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_traces.R
\name{average_traces}
\alias{average_traces}
\title{Averages ERG recordings from ERG lsits}
\usage{
average_traces(ERG_LIST, TFLASH)
}
\arguments{
\item{ERG_LIST}{ERG lists as generated by import_ERG}

\item{TFLASH}{the time of the flash stimulus}
}
\description{
This function averages singel ERG recordings from erg lists as generated by import_ERG. It returns a new ERG_LIST object with the additional items AVG, containing a tata.table with the columns Mean, SD, P25 and P75, and graphs, containing graphs of the individual recordings. If recording is from a single flash protocol, it removes offset based on the pre-stimulus$TRACES.
}
