% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileFromFolder.R
\name{compileFromFolder}
\alias{compileFromFolder}
\title{Compilres ERG recordings from multiple mice}
\usage{
compileFromFolder(RECS, Flicker = "Flicker")
}
\arguments{
\item{RECS}{a list of files containing ERG recordins}
}
\description{
This function imports ERG recordings from files in given folder and compiles them into the variables TRACES, STATS and STIMTYPES of a SUMMARY object
}
