% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropDulicateRecs.R
\name{dropDulicateRecs}
\alias{dropDulicateRecs}
\title{drops Recordings from ERG_LIST objects}
\usage{
dropDulicateRecs(ERG_LIST)
}
\arguments{
\item{ERG_LIST}{an ERG_LIST object}

\item{row_name}{Row name of record to be dropped as in ERG_LIST$FEATURES}
}
\description{
This function drops duplicate ERG recordings ERG_list objects
}
