% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropRec.R
\name{dropRec}
\alias{dropRec}
\title{drops Recordings from ERG_LIST objects}
\usage{
dropRec(ERG_LIST, row_name)
}
\arguments{
\item{ERG_LIST}{an ERG_LIST object}

\item{row_name}{Row name of record to be dropped as in ERG_LIST$FEATURES}
}
\description{
This function imports ERG recordings from abf files in a given folder into one list-type element
}
