% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeasures.R
\name{getMeasures}
\alias{getMeasures}
\title{drops Recordings from ERG_LIST objects}
\usage{
getMeasures(ERG_LIST, TMAXBWAVE)
}
\arguments{
\item{ERG_LIST}{an ERG_LIST object}

\item{TMAXBWAVE}{Maximum time of B-Wave Peak}
}
\description{
This function measures Flash and flicker ERG recordings as stored in an ERG_List object. Measures are stored in data frames in the MEASURES slot. For Flicker stimuli, Averaces from all measures is tored in FLICKERSTATS.
}
