% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualMeasures.R
\name{manualMeasures}
\alias{manualMeasures}
\title{drops Recordings from ERG_LIST objects}
\usage{
manualMeasures(ERG_LIST)
}
\arguments{
\item{ERG_LIST}{an ERG_LIST object}

\item{TMAXBWAVE}{Maximum time of B-Wave Peak}
}
\description{
This function allows for manual adjustment of Flash ERG measured as stored in an ERG_List object. Measures are stored in data frames in the MEASURES slot and overwrite automacical measurements.
}
