% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ERG.R
\name{plot_ERG}
\alias{plot_ERG}
\title{Plots ERGs}
\usage{
plot_ERG(curr, title, samp, marks = NULL)
}
\arguments{
\item{curr}{TBC}

\item{title}{TBC}

\item{samp}{TBC}

\item{marks}{TBC}
}
\description{
TBC
}
