% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ERG.R
\name{plot_ERG}
\alias{plot_ERG}
\title{Plots ERGs}
\usage{
plot_ERG(curr, title, samp, marks = NULL)
}
\arguments{
\item{curr}{The current ERG trace to plot}

\item{title}{The tilte for the plot (eg. recording conditions)}

\item{samp}{The sampling rate of curr}

\item{marks}{Any markers to be shown, e.g a-wave and b-wave positions}
}
\description{
TBC
}
